* Hardification feature supported (to be tested) * Ontology cleaning : remove empty reccords
git-svn-id: http://vcs.taotesting.com/svn/generis/branches/GenerisMinimalApi@2883 ab00eea3-224b-49ca-824e-92964491e928

+++ core/kernel/persistence/hardapi/class.TableManager.php
@@ -9,7 +9,7 @@ error_reporting(E_ALL);
  *
  * This file is part of Generis Object Oriented API.
  *
- * Automatically generated on 05.05.2011, 10:10:07 with ArgoUML PHP module 
+ * Automatically generated on 05.05.2011, 10:10:07 with ArgoUML PHP module
  * (last revised $Date: 2010-01-12 20:14:42 +0100 (Tue, 12 Jan 2010) $)
  *
  * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
@@ -18,7 +18,7 @@ error_reporting(E_ALL);
  */
 
 if (0 > version_compare(PHP_VERSION, '5')) {
-    die('This file was generated for PHP 5');
+	die('This file was generated for PHP 5');
 }
 
 /* user defined includes */
@@ -39,42 +39,42 @@ if (0 > version_compare(PHP_VERSION, '5')) {
  */
 class core_kernel_persistence_hardapi_TableManager
 {
-    // --- ASSOCIATIONS ---
+	// --- ASSOCIATIONS ---
 
 
-    // --- ATTRIBUTES ---
+	// --- ATTRIBUTES ---
 
-    /**
-     * Short description of attribute name
-     *
-     * @access protected
-     * @var string
-     */
-    protected $name = '';
+	/**
+	 * Short description of attribute name
+	 *
+	 * @access protected
+	 * @var string
+	 */
+	protected $name = '';
 
-    /**
-     * Short description of attribute _tables
-     *
-     * @access private
-     * @var array
-     */
-    private static $_tables = array();
+	/**
+	 * Short description of attribute _tables
+	 *
+	 * @access private
+	 * @var array
+	 */
+	private static $_tables = array();
 
-    // --- OPERATIONS ---
+	// --- OPERATIONS ---
 
-    /**
-     * Short description of method __construct
-     *
-     * @access public
-     * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
-     * @param  string name
-     * @return mixed
-     */
-    public function __construct($name)
-    {
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AA begin
-        
-    	if(count(self::$_tables) == 0){
+	/**
+	 * Short description of method __construct
+	 *
+	 * @access public
+	 * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
+	 * @param  string name
+	 * @return mixed
+	 */
+	public function __construct($name)
+	{
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AA begin
+
+		if(count(self::$_tables) == 0){
 			$dbWrapper = core_kernel_classes_DbWrapper::singleton();
 			$result = $dbWrapper->execSql('SELECT DISTINCT `table` FROM class_to_table');
 			while(!$result->EOF){
@@ -86,47 +86,47 @@ class core_kernel_persistence_hardapi_TableManager
 			throw new core_kernel_persistence_hardapi_Exception("Dangerous table name '$name' . Table name must begin by a underscore, followed  with only alphanumeric, - and _ characters are allowed");
 		}
 		$this->name = $name;
-                
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AA end
-    }
+		
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AA end
+	}
 
-    /**
-     * Short description of method exists
-     *
-     * @access public
-     * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
-     * @return boolean
-     */
-    public function exists()
-    {
-        $returnValue = (bool) false;
+	/**
+	 * Short description of method exists
+	 *
+	 * @access public
+	 * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
+	 * @return boolean
+	 */
+	public function exists()
+	{
+		$returnValue = (bool) false;
 
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AD begin
-        
-        $returnValue = in_array($this->name, self::$_tables);
-        
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AD end
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AD begin
 
-        return (bool) $returnValue;
-    }
+		$returnValue = in_array($this->name, self::$_tables);
 
-    /**
-     * Short description of method create
-     *
-     * @access public
-     * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
-     * @param  array columns
-     * @return boolean
-     */
-    public function create($columns = array())
-    {
-        $returnValue = (bool) false;
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AD end
 
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AF begin
-        
-    	if(!$this->exists() && !empty($this->name)){
+		return (bool) $returnValue;
+	}
+
+	/**
+	 * Short description of method create
+	 *
+	 * @access public
+	 * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
+	 * @param  array columns
+	 * @return boolean
+	 */
+	public function create($columns = array())
+	{
+		$returnValue = (bool) false;
+
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AF begin
+		
+		if(!$this->exists() && !empty($this->name)){
 			$dbWrapper = core_kernel_classes_DbWrapper::singleton();
-			
+				
 			//build the query to create the main table
 			$query = "CREATE TABLE {$this->name} (
 						id int NOT NULL AUTO_INCREMENT,
@@ -141,7 +141,7 @@ class core_kernel_persistence_hardapi_TableManager
 					$query .= ", {$column['name']}";
 					if(isset($column['foreign']) && !empty($column['foreign'])){
 						$query .= " VARCHAR(255),";
-						$query .= " CONSTRAINT fk_{$column['name']} 
+						$query .= " CONSTRAINT fk_{$column['name']}
 									FOREIGN KEY ({$column['name']}) 
 									REFERENCES {$column['foreign']}(uri)
 								";
@@ -161,7 +161,7 @@ class core_kernel_persistence_hardapi_TableManager
 				//the user may not have the right to create a table
 				throw new core_kernel_persistence_hardapi_Exception("Unable to create the table {$this->name} : " .$dbWrapper->dbConnector->errorMsg());
 			}
-			
+				
 			//create the multi prop table if needed
 			$query = "CREATE TABLE {$this->name}Props (
 				id int NOT NULL AUTO_INCREMENT,
@@ -177,7 +177,7 @@ class core_kernel_persistence_hardapi_TableManager
 						FOREIGN KEY (instance_id) 
 						REFERENCES {$this->name}(id)
 			)DEFAULT CHARSET=utf8";
-			
+				
 			$dbWrapper->execSql($query);
 			if($dbWrapper->dbConnector->errorNo() !== 0){
 				if(DEBUG_MODE){
@@ -186,55 +186,55 @@ class core_kernel_persistence_hardapi_TableManager
 				throw new core_kernel_persistence_hardapi_Exception("Unable to create the table {$this->name}Props : " .$dbWrapper->dbConnector->errorMsg());
 			}
 			self::$_tables[] = "{$this->name}Props";
-			
+				
 			//auto reference
 			self::$_tables[] = $this->name;
 			$returnValue = true;
-			
+				
 		}
-        
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AF end
+		
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015AF end
 
-        return (bool) $returnValue;
-    }
+		return (bool) $returnValue;
+	}
 
-    /**
-     * Short description of method remove
-     *
-     * @access public
-     * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
-     * @return boolean
-     */
-    public function remove()
-    {
-        $returnValue = (bool) false;
+	/**
+	 * Short description of method remove
+	 *
+	 * @access public
+	 * @author Cedric Alfonsi, <cedric.alfonsi@tudor.lu>
+	 * @return boolean
+	 */
+	public function remove()
+	{
+		$returnValue = (bool) false;
 
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015B9 begin
-        
-    	if($this->exists() && !empty($this->name)){
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015B9 begin
+
+		if($this->exists() && !empty($this->name)){
 			$dbWrapper = core_kernel_classes_DbWrapper::singleton();
 
-			//remove the multi property table
-			if(in_array("{$this->name}Props", self::$_tables)){
-				$dbWrapper->execSql("DROP TABLE `{$this->name}Props`");
-				$tblKey = array_search("{$this->name}Props", self::$_tables);
-					if($tblKey !== false){
-						unset(self::$_tables[$tblKey]);
-					}
+			//remove the multi properties table
+			$dbWrapper->execSql("DROP TABLE `{$this->name}Props`");
+			if($dbWrapper->dbConnector->errorNo() !== 0){
+				throw new core_kernel_persistence_hardapi_Exception("Unable to remove the properties table {$this->name}Props :" .$dbWrapper->dbConnector->errorMsg());
 			}
-			
+			$tblKey = array_search("{$this->name}Props", self::$_tables);
+			if($tblKey !== false){
+				unset(self::$_tables[$tblKey]);
+			}
+				
 			//remove the table
 			$dbWrapper->execSql("DROP TABLE `{$this->name}`");
-			
 			if($dbWrapper->dbConnector->errorNo() === 0){
-				
+
 				//remove all the references to the table
-				$cascadeDelete = "DELETE class_to_table.*, resource_has_class.*, resource_to_table.* FROM class_to_table 
+				$cascadeDelete = "DELETE class_to_table.*, resource_has_class.*, resource_to_table.* FROM class_to_table
 									LEFT JOIN resource_has_class ON resource_has_class.class_id = class_to_table.id
 									LEFT JOIN resource_to_table ON resource_has_class.resource_id = resource_to_table.id
 									WHERE class_to_table.`table` = '{$this->name}' OR resource_to_table.`table` = '{$this->name}'";
 				$dbWrapper->execSql($cascadeDelete);
-				
+
 				if($dbWrapper->dbConnector->errorNo() === 0){
 
 					$tblKey = array_search($this->name, self::$_tables);
@@ -252,12 +252,12 @@ class core_kernel_persistence_hardapi_TableManager
 				//the user may not have the right to drop a table
 				throw new core_kernel_persistence_hardapi_Exception("Unable to remove the table {$this->name} : " .$dbWrapper->dbConnector->errorMsg());
 			}
-    	}
-        
-        // section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015B9 end
+		}
 
-        return (bool) $returnValue;
-    }
+		// section 127-0-1-1--5a63b0fb:12f72879be9:-8000:00000000000015B9 end
+
+		return (bool) $returnValue;
+	}
 
 } /* end of class core_kernel_persistence_hardapi_TableManager */
